EPMRPP-91885 || Code Review fix - 3
