+++ app/src/controllers/administrate/constants.js
@@ -16,4 +16,3 @@
 
 export const NAMESPACE = 'admin';
 export const FETCH_PROJECT_DATA = 'fetchProjectData';
-export const FETCH_ACTIVE_PROJECT_DATA = 'fetchActiveProjectData';
